lexer: still wip
Change the lexer that is only returns the string and the type of the
token should indicate what is could be.

But there is a lot of ambiguity in lexing, e.g. if its a bare key or
integer depends on the "location" of lexing, i.e. if it is before or
after a "=". This grammatical information is not available in a simple
lexer but in the parser. I think the best approach is what I have a
already and parser-driven lexer.